JBIDE-15152 jQuery Mobile palette contains mix of jQuery Mobile and HTML5 items
HTML entries are moved from jQuery Mobile to new category HTML
